move minio client
