+++ src/five_safes_tes_workbench/core/minio/minio_client.py
@@ -4,16 +4,16 @@ from pathlib import Path
 
 from minio import Minio
 
-from ..helpers.auth import resolve_bearer
-from ..helpers.minio import (
+from ...helpers.auth import resolve_bearer
+from ...helpers.minio import (
     download_result,
     exchange_minio_token,
     list_results,
 )
-from ..helpers.url import is_https, strip_scheme
-from ..schema.auth_schema import AuthValidationModel
-from ..schema.config_schema import ConfigValidationModel
-from ..utils.logger import get_logger
+from ...helpers.url import is_https, strip_scheme
+from ...schema.auth_schema import AuthValidationModel
+from ...schema.config_schema import ConfigValidationModel
+from ...utils.logger import get_logger
 
 logger = get_logger(__name__)
